[issue-1270] Remove unused packages & check TODO comments in project
