+++ package.json
@@ -51,7 +51,6 @@
     "@react-native-community/segmented-control": "^2.2.2",
     "@react-native-community/slider": "^4.4.2",
     "@react-native-masked-view/masked-view": "^0.2.9",
-    "@react-native-picker/picker": "^2.4.8",
     "@react-navigation/bottom-tabs": "^6.5.3",
     "@react-navigation/drawer": "^6.6.2",
     "@react-navigation/material-top-tabs": "^6.6.3",
@@ -64,9 +63,8 @@
     "@subwallet/react-ui": "^5.1.2-b72",
     "@subwallet/ui-keyring": "^0.1.1",
     "@walletconnect/utils": "2.8.4",
-    "boring-avatars": "^1.7.0",
+    "appcenter": "^5.0.0",
     "deprecated-react-native-prop-types": "^4.0.0",
-    "eslint-plugin-ft-flow": "^2.0.3",
     "eventemitter3": "^5.0.0",
     "humanize-duration": "^3.31.0",
     "i18next": "^22.4.9",
@@ -83,12 +81,9 @@
     "react-i18next": "^12.1.4",
     "react-native": "^0.72.4",
     "react-native-animatable": "^1.3.3",
-    "react-native-bcrypt": "^2.4.0",
-    "react-native-boring-avatars": "^1.0.1",
     "react-native-camera": "^4.2.1",
     "react-native-code-push": "^8.1.1",
     "react-native-config": "^1.5.1",
-    "react-native-confirmation-code-field": "^7.3.1",
     "react-native-device-info": "^10.3.0",
     "react-native-document-picker": "^8.1.3",
     "react-native-dynamic-tab-view": "^1.0.7-0",
@@ -109,7 +104,6 @@
     "react-native-navigation-bar-color": "^2.0.2",
     "react-native-pager-view": "^6.2.0",
     "react-native-permissions": "^3.10.1",
-    "react-native-picker-select": "^8.0.4",
     "react-native-progress": "^5.0.0",
     "react-native-qrcode-scanner": "^1.5.5",
     "react-native-qrcode-styled": "^0.2.1",
@@ -127,7 +121,6 @@
     "react-native-toast-notifications": "^3.3.1",
     "react-native-url-polyfill": "^2.0.0",
     "react-native-vector-icons": "^10.0.0",
-    "react-native-version-number": "^0.3.6",
     "react-native-video": "^5.2.1",
     "react-native-view-shot": "^3.5.0",
     "react-native-webview": "^13.6.2",
