+++ src/components/FlatListScreen.tsx
@@ -13,7 +13,6 @@ import { useFilterModal } from 'hooks/useFilterModal';
 import { SectionListData } from 'react-native/Libraries/Lists/SectionList';
 import { LazySectionList, SectionItem } from 'components/LazySectionList';
 
-//TODO: split FlatList in FlatListScreen to new component, use ImperativeHandle to setPageNumber
 export interface RightIconOpt {
   icon?: (iconProps: IconProps) => JSX.Element;
   title?: string;
