+++ src/screens/Account/RestoreJson/index.tsx
@@ -272,10 +272,12 @@ export const RestoreJson = () => {
               {accountsInfo.length > 1 ? (
                 <SelectItem
                   leftItemIcon={<AvatarGroup addresses={addresses} />}
-                  label={i18n.formatString(
-                    i18n.importAccount.importAccounts,
-                    String(accountsInfo.length).padStart(2, '0'),
-                  )}
+                  label={
+                    i18n.formatString(
+                      i18n.importAccount.importAccounts,
+                      String(accountsInfo.length).padStart(2, '0'),
+                    ) as string
+                  }
                   onPress={openModal}
                   rightIcon={<Icon phosphorIcon={DotsThree} size="sm" />}
                 />
