+++ src/screens/Transaction/SendFund/index.tsx
@@ -1,55 +1,84 @@
-import { _AssetRef, _ChainAsset, _ChainInfo, _MultiChainAsset } from '@subwallet/chain-list/types';
+// Copyright 2019-2022 @polkadot/extension-koni-ui authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+import { _AssetRef, _AssetType, _ChainAsset, _ChainInfo, _MultiChainAsset } from '@subwallet/chain-list/types';
+import { AssetSetting, ExtrinsicType } from '@subwallet/extension-base/background/KoniTypes';
+import { AccountJson } from '@subwallet/extension-base/background/types';
 import {
   _getAssetDecimals,
   _getOriginChainOfAsset,
   _isAssetFungibleToken,
   _isChainEvmCompatible,
+  _isTokenTransferredByEvm,
 } from '@subwallet/extension-base/services/chain-service/utils';
-import { AccountJson } from '@subwallet/extension-base/background/types';
-import { _ChainConnectionStatus, _ChainState } from '@subwallet/extension-base/services/chain-service/types';
-import { AssetSetting } from '@subwallet/extension-base/background/KoniTypes';
-import { TokenItemType, TokenSelector } from 'components/Modal/common/TokenSelector';
-import { findAccountByAddress } from 'utils/index';
-import { findNetworkJsonByGenesisHash } from 'utils/getNetworkJsonByGenesisHash';
-import { isAddress, isEthereumAddress } from '@polkadot/util-crypto';
-import { isAccountAll } from 'utils/accountAll';
+import { SWTransactionResponse } from '@subwallet/extension-base/services/transaction-service/types';
+import { addLazy, isSameAddress, removeLazy } from '@subwallet/extension-base/utils';
+import BigN from 'bignumber.js';
 import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react';
+
+import { BN, BN_ZERO } from '@polkadot/util';
+import { isAddress, isEthereumAddress } from '@polkadot/util-crypto';
+import { SendFundProps } from 'routes/transaction/transactionAction';
 import { useSelector } from 'react-redux';
 import { RootState } from 'stores/index';
-import useHandleSubmitTransaction from 'hooks/transaction/useHandleSubmitTransaction';
-import { useTransaction } from 'hooks/screen/Transaction/useTransaction';
-import { KeyboardAvoidingView, Platform, SafeAreaView, ScrollView, View } from 'react-native';
-import { NetworkField } from 'components/Field/Network';
-import { AccountSelectField } from 'components/Field/AccountSelect';
-import { TokenSelectField } from 'components/Field/TokenSelect';
-import { InputAmount } from 'components/Input/InputAmount';
-import { ChainInfo } from 'types/index';
-import { addLazy, isSameAddress } from '@subwallet/extension-base/utils';
-import BigN from 'bignumber.js';
 import { getMaxTransfer, makeCrossChainTransfer, makeTransfer, saveRecentAccountId } from 'messaging/index';
-import { Button, Icon } from 'components/design-system-ui';
-import { PaperPlaneTilt } from 'phosphor-react-native';
-import { FreeBalance } from 'screens/Transaction/parts/FreeBalance';
-import { SWTransactionResponse } from '@subwallet/extension-base/services/transaction-service/types';
-import { Warning } from 'components/Warning';
-import { ContainerHorizontalPadding, MarginBottomForSubmitButton } from 'styles/sharedStyles';
-import { RootStackParamList } from 'routes/index';
+import { findAccountByAddress } from 'utils/account';
+import { findNetworkJsonByGenesisHash } from 'utils/getNetworkJsonByGenesisHash';
+import { formatBalance } from 'utils/number';
+import useGetChainPrefixBySlug from 'hooks/chain/useGetChainPrefixBySlug';
+import { TokenItemType, TokenSelector } from 'components/Modal/common/TokenSelector';
+import useHandleSubmitTransaction from 'hooks/transaction/useHandleSubmitTransaction';
+import { isAccountAll } from 'utils/accountAll';
+import { ChainInfo, ChainItemType } from 'types/index';
 import { useSubWalletTheme } from 'hooks/useSubWalletTheme';
-import { useNavigation } from '@react-navigation/native';
-import { NativeStackNavigationProp } from '@react-navigation/native-stack';
+import { useToast } from 'react-native-toast-notifications';
+import usePreCheckAction from 'hooks/account/usePreCheckAction';
+import { TransactionFormValues, useTransaction } from 'hooks/screen/Transaction/useTransaction';
+import { useWatch } from 'react-hook-form';
+import {
+  Keyboard,
+  KeyboardAvoidingView,
+  Platform,
+  SafeAreaView,
+  ScrollView,
+  TouchableOpacity,
+  View,
+} from 'react-native';
 import { ScreenContainer } from 'components/ScreenContainer';
 import { Header } from 'components/Header';
 import { SubHeader } from 'components/SubHeader';
-import { BN_ZERO } from 'utils/chainBalances';
-import { formatBalance } from 'utils/number';
-import { useToast } from 'react-native-toast-notifications';
+import { useNavigation } from '@react-navigation/native';
+import { NativeStackNavigationProp } from '@react-navigation/native-stack';
+import { RootStackParamList } from 'routes/index';
+import { AccountSelectField } from 'components/Field/AccountSelect';
 import i18n from 'utils/i18n/i18n';
-import { SendFundProps } from 'routes/transaction/transactionAction';
-import { InputAddress } from 'components/Input/InputAddressV2';
-import useGetChainPrefixBySlug from 'hooks/chain/useGetChainPrefixBySlug';
-import { ModalRef } from 'types/modalRef';
+import { TokenSelectField } from 'components/Field/TokenSelect';
+import { InputAddress } from 'components/Input/InputAddress';
+import { NetworkField } from 'components/Field/Network';
+import { Button, Icon, Typography } from 'components/design-system-ui';
 import { AccountSelector } from 'components/Modal/common/AccountSelector';
 import { ChainSelector } from 'components/Modal/common/ChainSelector';
+import { FormItem } from 'components/common/FormItem';
+import { ValidateResult } from 'react-hook-form/dist/types/validator';
+import { Amount, isInvalidAmountValue } from 'screens/Transaction/SendFund/Amount';
+import { ArrowCircleRight, PaperPlaneRight, PaperPlaneTilt } from 'phosphor-react-native';
+import { getButtonIcon } from 'utils/button';
+import { UseControllerReturn } from 'react-hook-form/dist/types';
+import { AmountValueConverter } from 'screens/Transaction/SendFund/AmountValueConverter';
+import createStylesheet from './styles';
+import { useGetBalance } from 'hooks/balance';
+import { FreeBalanceDisplay } from 'screens/Transaction/parts/FreeBalanceDisplay';
+import { ModalRef } from 'types/modalRef';
+import useChainAssets from 'hooks/chain/useChainAssets';
+import { TransactionDone } from 'screens/Transaction/TransactionDone';
+
+interface TransferFormValues extends TransactionFormValues {
+  to: string;
+  destChain: string;
+  value: string;
+}
+
+type ViewStep = 1 | 2;
 
 function isAssetTypeValid(
   chainAsset: _ChainAsset,
@@ -63,7 +92,6 @@ function getTokenItems(
   address: string,
   accounts: AccountJson[],
   chainInfoMap: Record<string, _ChainInfo>,
-  chainStateMap: Record<string, _ChainState>,
   assetRegistry: Record<string, _ChainAsset>,
   assetSettingMap: Record<string, AssetSetting>,
   multiChainAssetMap: Record<string, _MultiChainAsset>,
@@ -75,7 +103,9 @@ function getTokenItems(
     return [];
   }
 
-  const ledgerNetwork = findNetworkJsonByGenesisHash(chainInfoMap, account.originGenesisHash)?.slug;
+  const isLedger = !!account.isHardware;
+  const validGen: string[] = account.availableGenesisHashes || [];
+  const validLedgerNetwork = validGen.map(genesisHash => findNetworkJsonByGenesisHash(chainInfoMap, genesisHash)?.slug);
   const isAccountEthereum = isEthereumAddress(address);
   const isSetTokenSlug = !!tokenGroupSlug && !!assetRegistry[tokenGroupSlug];
   const isSetMultiChainAssetSlug = !!tokenGroupSlug && !!multiChainAssetMap[tokenGroupSlug];
@@ -86,7 +116,7 @@ function getTokenItems(
     }
 
     const chainAsset = assetRegistry[tokenGroupSlug];
-    const isValidLedger = ledgerNetwork ? ledgerNetwork === chainAsset?.originChain : true;
+    const isValidLedger = isLedger ? isAccountEthereum || validLedgerNetwork.includes(chainAsset?.originChain) : true;
 
     if (isSetTokenSlug) {
       if (isAssetTypeValid(chainAsset, chainInfoMap, isAccountEthereum) && isValidLedger) {
@@ -109,7 +139,7 @@ function getTokenItems(
   const items: TokenItemType[] = [];
 
   Object.values(assetRegistry).forEach(chainAsset => {
-    const isValidLedger = ledgerNetwork ? ledgerNetwork === chainAsset.originChain : true;
+    const isValidLedger = isLedger ? isAccountEthereum || validLedgerNetwork.includes(chainAsset?.originChain) : true;
     const isTokenFungible = _isAssetFungibleToken(chainAsset);
 
     if (!(isTokenFungible && isAssetTypeValid(chainAsset, chainInfoMap, isAccountEthereum) && isValidLedger)) {
@@ -142,12 +172,12 @@ function getTokenAvailableDestinations(
   tokenSlug: string,
   xcmRefMap: Record<string, _AssetRef>,
   chainInfoMap: Record<string, _ChainInfo>,
-): ChainInfo[] {
+): ChainItemType[] {
   if (!tokenSlug) {
     return [];
   }
 
-  const result: ChainInfo[] = [];
+  const result: ChainItemType[] = [];
   const originChain = chainInfoMap[_getOriginChainOfAsset(tokenSlug)];
 
   // Firstly, push the originChain of token
@@ -168,6 +198,59 @@ function getTokenAvailableDestinations(
   });
 
   return result;
+}
+
+function getTokenItemsForViewStep2(
+  originTokenItems: TokenItemType[],
+  xcmRefMap: Record<string, _AssetRef>,
+  chainInfoMap: Record<string, _ChainInfo>,
+  senderAddress: string,
+  recipientAddress: string,
+) {
+  const isRecipientAddressEvmType = isEthereumAddress(recipientAddress);
+  const isFromAndToSameType = isEthereumAddress(senderAddress) === isEthereumAddress(recipientAddress);
+
+  return originTokenItems.filter(ti => {
+    if (isFromAndToSameType && _isChainEvmCompatible(chainInfoMap[ti.originChain]) === isRecipientAddressEvmType) {
+      return true;
+    }
+
+    for (let xKey in xcmRefMap) {
+      if (!xcmRefMap.hasOwnProperty(xKey)) {
+        continue;
+      }
+
+      const xcmRef = xcmRefMap[xKey];
+      if (xcmRef.srcAsset === ti.slug) {
+        const isDestChainEvmCompatible = _isChainEvmCompatible(chainInfoMap[xcmRef.destChain]);
+        if (isDestChainEvmCompatible === isRecipientAddressEvmType) {
+          return true;
+        }
+      }
+    }
+
+    return false;
+  });
+}
+
+function getTokenAvailableDestinationsForViewStep2(
+  originChainItems: ChainItemType[],
+  chainInfoMap: Record<string, _ChainInfo>,
+  senderAddress: string,
+  recipientAddress: string,
+  chainValue: string,
+) {
+  const isRecipientAddressEvmType = isEthereumAddress(recipientAddress);
+  const _isSameAddress = isSameAddress(senderAddress, recipientAddress);
+
+  return originChainItems.filter(ci => {
+    if (_isSameAddress && chainValue === ci.slug) {
+      return false;
+    }
+
+    const isDestChainEvmCompatible = _isChainEvmCompatible(chainInfoMap[ci.slug]);
+    return isDestChainEvmCompatible === isRecipientAddressEvmType;
+  });
 }
 
 const defaultFilterAccount = (account: AccountJson): boolean => !(isAccountAll(account.address) || account.isReadOnly);
@@ -204,15 +287,18 @@ const filterAccountFunc = (
   });
 
   return (account: AccountJson): boolean => {
-    const ledgerNetwork = findNetworkJsonByGenesisHash(chainInfoMap, account.originGenesisHash)?.slug;
+    const isLedger = !!account.isHardware;
     const isAccountEthereum = isEthereumAddress(account.address);
+    const validGen: string[] = account.availableGenesisHashes || [];
+    const validLedgerNetwork =
+      validGen.map(genesisHash => findNetworkJsonByGenesisHash(chainInfoMap, genesisHash)?.slug) || [];
 
     if (!defaultFilterAccount(account)) {
       return false;
     }
 
     return chainAssets.some(chainAsset => {
-      const isValidLedger = ledgerNetwork ? ledgerNetwork === chainAsset?.originChain : true;
+      const isValidLedger = isLedger ? isAccountEthereum || validLedgerNetwork.includes(chainAsset?.originChain) : true;
 
       return isAssetTypeValid(chainAsset, chainInfoMap, isAccountEthereum) && isValidLedger;
     });
@@ -225,315 +311,517 @@ export const SendFund = ({
   },
 }: SendFundProps) => {
   const theme = useSubWalletTheme().swThemes;
+  const stylesheet = createStylesheet(theme);
   const { show, hideAll } = useToast();
-  const chainInfoMap = useSelector((root: RootState) => root.chainStore.chainInfoMap);
-  const chainStateMap = useSelector((root: RootState) => root.chainStore.chainStateMap);
-  const assetRegistry = useSelector((root: RootState) => root.assetRegistry.assetRegistry);
-  const assetSettingMap = useSelector((root: RootState) => root.assetRegistry.assetSettingMap);
-  const multiChainAssetMap = useSelector((root: RootState) => root.assetRegistry.multiChainAssetMap);
-  const xcmRefMap = useSelector((root: RootState) => root.assetRegistry.xcmRefMap);
-  const accounts = useSelector((state: RootState) => state.accountState.accounts);
-  const isAllAccount = useSelector((state: RootState) => state.accountState.isAllAccount);
-  const [maxTransfer, setMaxTransfer] = useState<string>('0');
-  const [loading, setLoading] = useState(false);
-  const [isTransferAll, setIsTransferAll] = useState(false);
-  const [isBalanceReady, setIsBalanceReady] = useState(true);
-  const [forceUpdateMaxValue, setForceUpdateMaxValue] = useState<object | undefined>(undefined);
   const navigation = useNavigation<NativeStackNavigationProp<RootStackParamList>>();
-
-  const [isToAddressDirty, setToAddressDirty] = useState<boolean>(false);
+  const [viewStep, setViewStep] = useState<ViewStep>(1);
+  const [transactionDone, setTransactionDone] = useState<boolean>(false);
   const accountSelectorRef = useRef<ModalRef>();
   const tokenSelectorRef = useRef<ModalRef>();
   const chainSelectorRef = useRef<ModalRef>();
 
-  const handleTransferAll = useCallback((value: boolean) => {
-    setForceUpdateMaxValue({});
-    setIsTransferAll(value);
-  }, []);
+  const {
+    title,
+    form: {
+      setValue,
+      resetField,
+      getValues,
+      control,
+      handleSubmit,
+      trigger,
+      setFocus,
+      formState: { errors, dirtyFields },
+    },
+    onChangeFromValue: setFrom,
+    onChangeAssetValue: setAsset,
+    onChangeChainValue: setChain,
+    onTransactionDone: onDone,
+    transactionDoneInfo,
+    checkChainConnected,
+  } = useTransaction<TransferFormValues>('send-fund', {
+    mode: 'onBlur',
+    reValidateMode: 'onBlur',
+    defaultValues: {
+      destChain: '',
+      to: '',
+    },
+  });
 
-  const formConfig = {
-    to: {
-      name: i18n.inputLabel.sendTo,
-      value: '',
-    },
-    destChain: {
-      name: 'Destination chain',
-      value: '',
-    },
+  const {
+    asset: assetValue,
+    chain: chainValue,
+    from: fromValue,
+    to: toValue,
+    destChain: destChainValue,
+    value: transferAmount,
+  } = {
+    ...useWatch<TransferFormValues>({ control }),
+    ...getValues(),
   };
 
-  const { title, formState, onChangeValue, onUpdateErrors, onDone, onChangeFromValue, onChangeAssetValue } =
-    useTransaction('send-fund', formConfig, {});
-  const { asset, chain, destChain, from, to, value: amount } = formState.data;
-  const destChainNetworkPrefix = useGetChainPrefixBySlug(destChain);
-  const destChainGenesisHash = chainInfoMap[destChain]?.substrateInfo?.genesisHash || '';
-  const { onError, onSuccess } = useHandleSubmitTransaction(onDone, handleTransferAll);
-  const isDataNotReady = !to || !amount;
+  const {
+    error: isGetBalanceError,
+    isLoading: isGetBalanceLoading,
+    nativeTokenBalance,
+    nativeTokenSlug,
+    tokenBalance,
+  } = useGetBalance(chainValue, fromValue, assetValue);
+
+  const { chainInfoMap, chainStateMap } = useSelector((root: RootState) => root.chainStore);
+  const { assetSettingMap, multiChainAssetMap, xcmRefMap } = useSelector((root: RootState) => root.assetRegistry);
+  const assetRegistry = useChainAssets().chainAssetRegistry;
+  const { accounts, isAllAccount } = useSelector((state: RootState) => state.accountState);
+  const [maxTransfer, setMaxTransfer] = useState<string>('0');
+  const checkAction = usePreCheckAction(
+    fromValue,
+    true,
+    'The account you are using is {{accountTitle}}, you cannot send assets with it',
+  );
+
+  const [loading, setLoading] = useState(false);
+  const [isTransferAll, setIsTransferAll] = useState(false);
+  const [, update] = useState({});
+  const [isBalanceReady, setIsBalanceReady] = useState(true);
+  const [forceUpdateValue, setForceUpdateValue] = useState<{ value: string | null } | undefined>(undefined);
+  const chainStatus = useMemo(() => chainStateMap[chainValue]?.connectionStatus, [chainValue, chainStateMap]);
+
+  const senderAccountName = useMemo(() => {
+    if (!fromValue) {
+      return i18n.inputLabel.selectAcc;
+    }
+
+    const targetAccount = accounts.find(a => a.address === fromValue);
+
+    return targetAccount?.name || '';
+  }, [accounts, fromValue]);
+
+  const handleTransferAll = useCallback(
+    (value: boolean) => {
+      setForceUpdateValue({ value: maxTransfer });
+      setIsTransferAll(value);
+    },
+    [maxTransfer],
+  );
+
+  const triggerOnChangeValue = () => {
+    setForceUpdateValue({ value: transferAmount });
+  };
+
+  const { onError, onSuccess } = useHandleSubmitTransaction(
+    onDone,
+    setTransactionDone,
+    triggerOnChangeValue,
+    handleTransferAll,
+  );
 
   const accountItems = useMemo(() => {
     return accounts.filter(filterAccountFunc(chainInfoMap, assetRegistry, multiChainAssetMap, tokenGroupSlug));
   }, [accounts, assetRegistry, chainInfoMap, multiChainAssetMap, tokenGroupSlug]);
 
-  const destChainItems = useMemo<ChainInfo[]>(() => {
-    return getTokenAvailableDestinations(asset, xcmRefMap, chainInfoMap);
-  }, [chainInfoMap, asset, xcmRefMap]);
+  const destChainItems = useMemo<ChainItemType[]>(() => {
+    return getTokenAvailableDestinations(assetValue, xcmRefMap, chainInfoMap);
+  }, [assetValue, xcmRefMap, chainInfoMap]);
+
+  const destChainItemsViewStep2 = useMemo<ChainItemType[]>(() => {
+    if (viewStep !== 2) {
+      return [];
+    }
+
+    return getTokenAvailableDestinationsForViewStep2(destChainItems, chainInfoMap, fromValue, toValue, chainValue);
+  }, [viewStep, destChainItems, chainInfoMap, fromValue, toValue, chainValue]);
 
   const currentChainAsset = useMemo(() => {
-    return asset ? assetRegistry[asset] : undefined;
-  }, [assetRegistry, asset]);
+    return assetValue ? assetRegistry[assetValue] : undefined;
+  }, [assetValue, assetRegistry]);
 
   const decimals = useMemo(() => {
-    return currentChainAsset ? _getAssetDecimals(currentChainAsset) : -1;
+    return currentChainAsset ? _getAssetDecimals(currentChainAsset) : 0;
   }, [currentChainAsset]);
+
+  const extrinsicType = useMemo((): ExtrinsicType => {
+    if (!currentChainAsset) {
+      return ExtrinsicType.UNKNOWN;
+    } else {
+      if (chainValue !== destChainValue) {
+        return ExtrinsicType.TRANSFER_XCM;
+      } else {
+        if (currentChainAsset.assetType === _AssetType.NATIVE) {
+          return ExtrinsicType.TRANSFER_BALANCE;
+        } else {
+          return ExtrinsicType.TRANSFER_TOKEN;
+        }
+      }
+    }
+  }, [chainValue, currentChainAsset, destChainValue]);
+
+  // const fromChainNetworkPrefix = useGetChainPrefixBySlug(chainValue); // will use it for account selector later
+  const destChainNetworkPrefix = useGetChainPrefixBySlug(destChainValue);
+  const destChainGenesisHash = chainInfoMap[destChainValue]?.substrateInfo?.genesisHash || '';
 
   const tokenItems = useMemo<TokenItemType[]>(() => {
     return getTokenItems(
-      from,
+      fromValue,
       accounts,
       chainInfoMap,
-      chainStateMap,
       assetRegistry,
       assetSettingMap,
       multiChainAssetMap,
       tokenGroupSlug,
-    );
-  }, [accounts, assetRegistry, assetSettingMap, chainInfoMap, chainStateMap, from, multiChainAssetMap, tokenGroupSlug]);
-
-  const validateRecipientAddress = useCallback(
-    (_recipientAddress: string, _from: string, _chain: string, _destChain: string) => {
-      if (!_recipientAddress) {
-        onUpdateErrors('to')([i18n.errorMessage.recipientAddressIsRequired]);
-        return false;
-      }
-
-      if (!isAddress(_recipientAddress)) {
-        onUpdateErrors('to')([i18n.errorMessage.invalidRecipientAddress]);
-
-        return false;
-      }
-
-      if (!_from || !_chain || !_destChain) {
-        return false;
-      }
-
-      const isOnChain = _chain === _destChain;
-
-      if (isOnChain) {
-        if (isSameAddress(_from, _recipientAddress)) {
-          onUpdateErrors('to')([i18n.errorMessage.sameAddressError]);
-          return false;
-        }
-
-        const isNotSameAddressType =
-          (isEthereumAddress(_from) && !!_recipientAddress && !isEthereumAddress(_recipientAddress)) ||
-          (!isEthereumAddress(_from) && !!_recipientAddress && isEthereumAddress(_recipientAddress));
-
-        if (isNotSameAddressType) {
-          onUpdateErrors('to')([i18n.errorMessage.notSameAddressTypeError]);
-          return false;
+    ).sort((a, b) => {
+      if (checkChainConnected(a.originChain)) {
+        if (checkChainConnected(b.originChain)) {
+          return 0;
+        } else {
+          return -1;
         }
       } else {
-        const isDestChainEvmCompatible = _isChainEvmCompatible(chainInfoMap[_destChain]);
-
-        if (isDestChainEvmCompatible !== isEthereumAddress(_recipientAddress)) {
-          onUpdateErrors('to')([
-            i18n.formatString(
-              i18n.errorMessage.recipientAddressMustBeType,
-              isDestChainEvmCompatible ? 'EVM' : 'substrate',
-            ),
-          ]);
-          return false;
+        if (checkChainConnected(b.originChain)) {
+          return 1;
+        } else {
+          return 0;
         }
       }
+    });
+  }, [
+    accounts,
+    assetRegistry,
+    assetSettingMap,
+    chainInfoMap,
+    checkChainConnected,
+    fromValue,
+    multiChainAssetMap,
+    tokenGroupSlug,
+  ]);
 
-      onUpdateErrors('to')([]);
-      return true;
-    },
-    [chainInfoMap, onUpdateErrors],
+  const tokenItemsViewStep2 = useMemo(() => {
+    if (viewStep !== 2 || !fromValue || !toValue) {
+      return [];
+    }
+
+    return getTokenItemsForViewStep2(tokenItems, xcmRefMap, chainInfoMap, fromValue, toValue).sort((a, b) => {
+      if (checkChainConnected(a.originChain)) {
+        if (checkChainConnected(b.originChain)) {
+          return 0;
+        } else {
+          return -1;
+        }
+      } else {
+        if (checkChainConnected(b.originChain)) {
+          return 1;
+        } else {
+          return 0;
+        }
+      }
+    });
+  }, [chainInfoMap, checkChainConnected, fromValue, toValue, tokenItems, viewStep, xcmRefMap]);
+
+  const recipientAddressRules = useMemo(
+    () => ({
+      validate: (
+        _recipientAddress: string,
+        { chain, destChain, from }: TransactionFormValues,
+      ): Promise<ValidateResult> => {
+        if (!_recipientAddress) {
+          return Promise.resolve(i18n.errorMessage.recipientAddressIsRequired);
+        }
+
+        if (!isAddress(_recipientAddress)) {
+          return Promise.resolve(i18n.errorMessage.invalidRecipientAddress);
+        }
+
+        if (!from || !chain || !destChain) {
+          return Promise.resolve(undefined);
+        }
+
+        const isOnChain = chain === destChain;
+
+        const account = findAccountByAddress(accounts, _recipientAddress);
+
+        if (isOnChain) {
+          if (isSameAddress(from, _recipientAddress)) {
+            return Promise.resolve(i18n.errorMessage.sameAddressError);
+          }
+
+          const isNotSameAddressType =
+            (isEthereumAddress(from) && !!_recipientAddress && !isEthereumAddress(_recipientAddress)) ||
+            (!isEthereumAddress(from) && !!_recipientAddress && isEthereumAddress(_recipientAddress));
+
+          if (isNotSameAddressType) {
+            return Promise.resolve(i18n.errorMessage.notSameAddressTypeError);
+          }
+        } else {
+          const isDestChainEvmCompatible = _isChainEvmCompatible(chainInfoMap[destChain]);
+
+          if (isDestChainEvmCompatible !== isEthereumAddress(_recipientAddress)) {
+            return Promise.resolve(
+              i18n.formatString(
+                i18n.errorMessage.recipientAddressMustBeType,
+                isDestChainEvmCompatible ? 'EVM' : 'substrate',
+              ),
+            );
+          }
+        }
+
+        if (account?.isHardware) {
+          const destChainInfo = chainInfoMap[destChain];
+          const availableGen: string[] = account.availableGenesisHashes || [];
+
+          if (
+            !isEthereumAddress(account.address) &&
+            !availableGen.includes(destChainInfo?.substrateInfo?.genesisHash || '')
+          ) {
+            const destChainName = destChainInfo?.name || 'Unknown';
+
+            return Promise.resolve(
+              `'Wrong network. Your Ledger account is not supported by ${destChainName}. Please choose another receiving account and try again.'`,
+            );
+          }
+        }
+
+        return Promise.resolve(undefined);
+      },
+    }),
+    [accounts, chainInfoMap],
   );
 
-  const onChangeRecipientAddress = useCallback(
-    (currentTextValue: string) => {
-      setToAddressDirty(true);
-      onChangeValue('to')(currentTextValue);
-      if (currentTextValue !== to) {
-        validateRecipientAddress(currentTextValue, from, chain, destChain);
-      }
-    },
-    [chain, destChain, from, onChangeValue, to, validateRecipientAddress],
+  const amountRules = useMemo(
+    () => ({
+      validate: (amount: string): Promise<ValidateResult> => {
+        if (isInvalidAmountValue(amount)) {
+          return Promise.resolve(i18n.errorMessage.invalidAmount);
+        }
+
+        if (!amount) {
+          return Promise.resolve(i18n.errorMessage.amountRequiredError);
+        }
+
+        if (new BigN(amount).eq(new BigN(0))) {
+          return Promise.resolve(i18n.errorMessage.amountMustBeGreaterThanZero);
+        }
+
+        if (new BigN(amount).gt(new BigN(maxTransfer))) {
+          const maxString = formatBalance(maxTransfer, decimals);
+
+          return Promise.resolve(i18n.formatString(i18n.errorMessage.amountMustBeEqualOrLessThan, maxString));
+        }
+
+        return Promise.resolve(undefined);
+      },
+    }),
+    [decimals, maxTransfer],
   );
 
-  useEffect(() => {
-    if (scanRecipient) {
-      setToAddressDirty(true);
-      onChangeValue('to')(scanRecipient);
-      if (isAddress(scanRecipient)) {
-        saveRecentAccountId(scanRecipient).catch(console.error);
+  const _onChangeFrom = (item: AccountJson) => {
+    setFrom(item.address);
+    accountSelectorRef?.current?.onCloseModal();
+    resetField('asset');
+    setForceUpdateValue(undefined);
+    setIsTransferAll(false);
+  };
+
+  const _onChangeAsset = (item: TokenItemType) => {
+    setAsset(item.slug);
+    const currentDestChainItems = getTokenAvailableDestinationsForViewStep2(
+      destChainItems,
+      chainInfoMap,
+      fromValue,
+      toValue,
+      item.originChain,
+    );
+
+    if (viewStep === 2) {
+      if (currentDestChainItems.some(destChainItem => destChainItem.slug === item.originChain)) {
+        setValue('destChain', item.originChain);
+      } else {
+        setValue('destChain', '');
       }
-    }
-  }, [onChangeValue, scanRecipient]);
-
-  useEffect(() => {
-    if (scanRecipient === to) {
-      validateRecipientAddress(to, from, chain, destChain);
-    }
-  }, [chain, destChain, from, scanRecipient, to, validateRecipientAddress]);
-
-  const validateAmount = useCallback(
-    (_amount: string, _maxTransfer: string, isValidInput?: boolean) => {
-      if (isValidInput === false) {
-        onUpdateErrors('value')([i18n.errorMessage.invalidAmount]);
-
-        return false;
-      }
-
-      if (!_amount) {
-        onUpdateErrors('value')([i18n.errorMessage.amountRequiredError]);
-
-        return false;
-      }
-
-      if (new BigN(_amount).eq(new BigN(0))) {
-        onUpdateErrors('value')([i18n.errorMessage.amountMustBeGreaterThanZero]);
-
-        return false;
-      }
-
-      if (new BigN(_amount).gt(new BigN(_maxTransfer))) {
-        const maxString = formatBalance(_maxTransfer, decimals);
-        onUpdateErrors('value')([i18n.formatString(i18n.errorMessage.amountMustBeEqualOrLessThan, maxString)]);
-
-        return false;
-      }
-
-      onUpdateErrors('value')(undefined);
-      return true;
-    },
-    [decimals, onUpdateErrors],
-  );
-
-  const _onChangeAmount = useCallback(
-    (_amount: string, isValidInput: boolean) => {
-      addLazy(
-        'input-amount',
-        () => {
-          onChangeValue('value')(_amount);
-          validateAmount(_amount, maxTransfer, isValidInput);
-        },
-        1000,
-      );
-    },
-    [maxTransfer, onChangeValue, validateAmount],
-  );
-
-  const senderAccountName = useMemo(() => {
-    if (!from) {
-      return i18n.inputLabel.selectAcc;
-    }
-
-    const targetAccount = accounts.find(a => a.address === from);
-
-    return targetAccount?.name || '';
-  }, [accounts, from]);
-
-  const onSubmit = useCallback(() => {
-    if (chainStateMap[chain].connectionStatus === _ChainConnectionStatus.DISCONNECTED) {
-      show(`${destChain} ${i18n.errorMessage.networkDisconected}`, { type: 'danger' });
-      return;
-    }
-
-    setLoading(true);
-
-    const isAmountValid = validateAmount(amount, maxTransfer);
-    const isRecipientAddressValid = validateRecipientAddress(to, from, chain, destChain);
-
-    const isFormValid = isAmountValid && isRecipientAddressValid;
-
-    if (!isFormValid) {
-      setLoading(false);
-      return;
-    }
-
-    let sendPromise: Promise<SWTransactionResponse>;
-
-    if (chain === destChain) {
-      // Transfer token or send fund
-      sendPromise = makeTransfer({
-        from,
-        networkKey: chain,
-        to,
-        tokenSlug: asset,
-        value: amount,
-        transferAll: isTransferAll,
-      });
     } else {
-      const acc = findAccountByAddress(accounts, from);
+      setValue('destChain', item.originChain);
+    }
 
-      if (acc?.isHardware) {
+    tokenSelectorRef?.current?.onCloseModal();
+    setForceUpdateValue(undefined);
+    setIsTransferAll(false);
+  };
+
+  const _onChangeDestChain = (item: ChainInfo) => {
+    setValue('destChain', item.slug);
+    chainSelectorRef?.current?.onCloseModal();
+  };
+
+  const onSubheaderPressBack = useCallback(() => {
+    if (viewStep === 1) {
+      navigation.goBack();
+    } else {
+      setViewStep(1);
+      resetField('value', {
+        keepDirty: false,
+        keepError: false,
+        keepTouched: false,
+      });
+      setForceUpdateValue({ value: null });
+      setIsTransferAll(false);
+    }
+  }, [navigation, resetField, viewStep]);
+
+  // Submit transaction
+  const onSubmit = useCallback(
+    (values: TransferFormValues) => {
+      Keyboard.dismiss();
+      setLoading(true);
+      const { chain, destChain, to, value, from, asset } = values;
+
+      let sendPromise: Promise<SWTransactionResponse>;
+
+      const account = findAccountByAddress(accounts, from);
+
+      if (!account) {
         setLoading(false);
         hideAll();
-        show('This feature is not available for Ledger account', { type: 'danger' });
+        show("Can't find account");
 
         return;
       }
 
-      // Make cross chain transfer
-      sendPromise = makeCrossChainTransfer({
-        destinationNetworkKey: destChain,
-        from,
-        originNetworkKey: chain,
-        tokenSlug: asset,
-        to,
-        value: amount,
-        transferAll: isTransferAll,
-      });
+      const isLedger = !!account.isHardware;
+      const isEthereum = isEthereumAddress(account.address);
+      const chainAsset = assetRegistry[asset];
+
+      if (chain === destChain) {
+        if (isLedger) {
+          if (isEthereum) {
+            if (!_isTokenTransferredByEvm(chainAsset)) {
+              setLoading(false);
+              hideAll();
+              show('Ledger does not support transfer for this token');
+
+              return;
+            }
+          }
+        }
+
+        // Transfer token or send fund
+        sendPromise = makeTransfer({
+          from,
+          networkKey: chain,
+          to: to,
+          tokenSlug: asset,
+          value: value,
+          transferAll: isTransferAll,
+        });
+      } else {
+        if (isLedger) {
+          setLoading(false);
+          hideAll();
+          show('This feature is not available for Ledger account');
+
+          return;
+        }
+
+        // Make cross chain transfer
+        sendPromise = makeCrossChainTransfer({
+          destinationNetworkKey: destChain,
+          from,
+          originNetworkKey: chain,
+          tokenSlug: asset,
+          to,
+          value,
+          transferAll: isTransferAll,
+        });
+      }
+
+      setTimeout(() => {
+        // Handle transfer action
+        sendPromise
+          .then(onSuccess)
+          .catch(onError)
+          .finally(() => {
+            setLoading(false);
+          });
+      }, 300);
+    },
+    [accounts, assetRegistry, isTransferAll, hideAll, show, onSuccess, onError],
+  );
+
+  const isNextButtonDisable = (() => {
+    if (!isBalanceReady || !fromValue || !assetValue || !destChainValue || !toValue) {
+      return true;
     }
 
-    setTimeout(() => {
-      // Handle transfer action
-      sendPromise
-        .then(onSuccess)
-        .catch(onError)
-        .finally(() => {
-          setLoading(false);
-        });
-    }, 300);
-  }, [
-    chainStateMap,
-    destChain,
-    validateAmount,
-    amount,
-    maxTransfer,
-    validateRecipientAddress,
-    to,
-    from,
-    chain,
-    show,
-    asset,
-    isTransferAll,
-    accounts,
-    hideAll,
-    onSuccess,
-    onError,
-  ]);
+    return !!errors.to;
+  })();
 
-  const onSetMaxTransferable = useCallback(
-    (value: boolean) => {
-      const bnMaxTransfer = new BigN(maxTransfer);
+  const isSubmitButtonDisable = (() => {
+    return loading || isNextButtonDisable || !transferAmount || !!errors.value;
+  })();
 
-      if (!bnMaxTransfer.isZero()) {
-        setIsTransferAll(value);
-      }
+  const onInputChangeAmount = useCallback(() => {
+    setIsTransferAll(false);
+  }, []);
+
+  const reValidate = () => trigger('to');
+
+  const renderAmountInput = useCallback(
+    ({ field: { onBlur, onChange, value, ref } }: UseControllerReturn<TransferFormValues>) => {
+      return (
+        <>
+          <Amount
+            ref={ref}
+            value={value}
+            forceUpdateValue={forceUpdateValue}
+            onChangeValue={onChange}
+            onInputChange={onInputChangeAmount}
+            onBlur={onBlur}
+            onSideEffectChange={onBlur}
+            decimals={decimals}
+            placeholder={'0'}
+            showMaxButton
+          />
+          <AmountValueConverter
+            value={isInvalidAmountValue(value) ? '0' : value || '0'}
+            tokenSlug={assetValue}
+            style={stylesheet.amountValueConverter}
+          />
+        </>
+      );
     },
-    [maxTransfer],
+    [assetValue, decimals, forceUpdateValue, onInputChangeAmount, stylesheet.amountValueConverter],
   );
 
   useEffect(() => {
+    setIsBalanceReady(!isGetBalanceLoading && !isGetBalanceError);
+  }, [isGetBalanceError, isGetBalanceLoading]);
+
+  useEffect(() => {
+    if (scanRecipient) {
+      setValue('to', scanRecipient, {
+        shouldDirty: true,
+        shouldTouch: true,
+      });
+    }
+  }, [scanRecipient, setValue]);
+
+  useEffect(() => {
+    setForceUpdateValue(isTransferAll ? { value: maxTransfer } : undefined);
+  }, [destChainValue, isTransferAll, maxTransfer]);
+
+  useEffect(() => {
+    const { asset, from } = getValues();
+
+    const updateInfoWithTokenSlug = (tokenSlug: string) => {
+      const tokenInfo = assetRegistry[tokenSlug];
+
+      setValue('asset', tokenSlug);
+      setValue('chain', tokenInfo.originChain);
+      setValue('destChain', tokenInfo.originChain);
+      setChain(tokenInfo.originChain);
+    };
+
     if (tokenItems.length) {
+      let isApplyDefaultAsset = true;
+
       if (!asset) {
         const account = findAccountByAddress(accounts, from);
-
-        let pass = false;
 
         if (account?.originGenesisHash) {
           const network = findNetworkJsonByGenesisHash(chainInfoMap, account.originGenesisHash);
@@ -542,42 +830,33 @@ export const SendFund = ({
             const token = tokenItems.find(item => item.originChain === network.slug);
 
             if (token) {
-              onChangeAssetValue(tokenItems[0].slug);
-              onChangeValue('chain')(assetRegistry[token.slug].originChain);
-              onChangeValue('destChain')(assetRegistry[token.slug].originChain);
-              pass = true;
+              updateInfoWithTokenSlug(token.slug);
+              isApplyDefaultAsset = false;
             }
           }
         }
-
-        if (!pass) {
-          onChangeAssetValue(tokenItems[0].slug);
-          onChangeValue('chain')(assetRegistry[tokenItems[0].slug].originChain);
-          onChangeValue('destChain')(assetRegistry[tokenItems[0].slug].originChain);
-        }
       } else {
-        const isSelectedTokenInList = tokenItems.some(i => i.slug === asset);
+        // Apply default asset if current asset is not in token list
+        isApplyDefaultAsset = !tokenItems.some(i => i.slug === asset);
+      }
 
-        if (!isSelectedTokenInList) {
-          onChangeAssetValue(tokenItems[0].slug);
-          onChangeValue('chain')(assetRegistry[tokenItems[0].slug].originChain);
-          onChangeValue('destChain')(assetRegistry[tokenItems[0].slug].originChain);
-        }
+      if (isApplyDefaultAsset) {
+        updateInfoWithTokenSlug(tokenItems[0].slug);
       }
     }
-  }, [accounts, tokenItems, assetRegistry, chainInfoMap, asset, from, onChangeValue, onChangeAssetValue]);
+  }, [accounts, tokenItems, assetRegistry, setChain, chainInfoMap, getValues, setValue]);
 
   // Get max transfer value
   useEffect(() => {
     let cancel = false;
 
-    if (from && asset) {
+    if (fromValue && assetValue) {
       getMaxTransfer({
-        address: from,
-        networkKey: assetRegistry[asset].originChain,
-        token: asset,
-        isXcmTransfer: chain !== destChain,
-        destChain,
+        address: fromValue,
+        networkKey: assetRegistry[assetValue].originChain,
+        token: assetValue,
+        isXcmTransfer: chainValue !== destChainValue,
+        destChain: destChainValue,
       })
         .then(balance => {
           !cancel && setMaxTransfer(balance.value);
@@ -587,9 +866,11 @@ export const SendFund = ({
         })
         .finally(() => {
           if (!cancel) {
-            if (amount) {
+            const value = getValues('value');
+
+            if (value) {
               setTimeout(() => {
-                validateAmount(amount, maxTransfer);
+                trigger('value').finally(() => update({}));
               }, 100);
             }
           }
@@ -599,179 +880,260 @@ export const SendFund = ({
     return () => {
       cancel = true;
     };
-  }, [amount, asset, assetRegistry, assetSettingMap, chain, destChain, from, maxTransfer, validateAmount]);
+  }, [assetValue, assetRegistry, chainStatus, chainValue, destChainValue, fromValue, getValues, trigger]);
 
   useEffect(() => {
-    const bnTransferAmount = new BigN(amount || '0');
-    const bnMaxTransfer = new BigN(maxTransfer || '0');
+    const bnTransferAmount = new BN(isInvalidAmountValue(transferAmount) ? '0' : transferAmount || '0');
+    const bnMaxTransfer = new BN(maxTransfer || '0');
 
     if (bnTransferAmount.gt(BN_ZERO) && bnTransferAmount.eq(bnMaxTransfer)) {
       setIsTransferAll(true);
     }
-  }, [maxTransfer, amount]);
+  }, [maxTransfer, transferAmount]);
 
-  const buttonIcon = useCallback((color: string) => {
-    return <Icon phosphorIcon={PaperPlaneTilt} weight={'fill'} size={'lg'} iconColor={color} />;
-  }, []);
+  useEffect(() => {
+    if (scanRecipient) {
+      if (isAddress(scanRecipient)) {
+        saveRecentAccountId(scanRecipient).catch(console.error);
+      }
+    }
+  }, [scanRecipient]);
+
+  useEffect(() => {
+    if (fromValue && chainValue && destChainValue && dirtyFields.to) {
+      addLazy(
+        'trigger-validate-send-fund-to',
+        () => {
+          trigger('to');
+        },
+        100,
+      );
+    }
+
+    return () => {
+      removeLazy('trigger-validate-send-fund-to');
+    };
+  }, [chainValue, destChainValue, dirtyFields.to, fromValue, trigger]);
+
+  useEffect(() => {
+    addLazy('auto-focus-send-fund', () => {
+      if (viewStep === 2) {
+        setFocus('value');
+      }
+    });
+
+    return () => {
+      removeLazy('auto-focus-send-fund');
+    };
+  }, [setFocus, viewStep]);
 
   return (
-    <KeyboardAvoidingView behavior={Platform.OS === 'ios' ? 'padding' : undefined} style={{ flex: 1 }}>
-      <ScreenContainer>
-        <>
-          <Header disabled={loading} />
+    <>
+      {!transactionDone ? (
+        <KeyboardAvoidingView behavior={Platform.OS === 'ios' ? 'padding' : 'height'} style={{ flex: 1 }}>
+          <ScreenContainer>
+            <>
+              <Header disabled={loading} />
 
-          <View style={{ paddingTop: 16 }}>
-            <SubHeader
-              title={title}
-              onPressBack={() => navigation.goBack()}
-              disabled={loading}
-              titleTextAlign={'left'}
-            />
-          </View>
-
-          <>
-            <ScrollView
-              style={{ ...ContainerHorizontalPadding, paddingTop: theme.size, flex: 1 }}
-              keyboardShouldPersistTaps={'handled'}>
-              {isAllAccount && (
-                <AccountSelector
-                  items={accountItems}
-                  selectedValueMap={{ [from]: true }}
-                  onSelectItem={item => {
-                    onChangeFromValue(item.address);
-                    setForceUpdateMaxValue(undefined);
-                    accountSelectorRef && accountSelectorRef.current?.onCloseModal();
-                    setIsTransferAll(false);
-                    isToAddressDirty && validateRecipientAddress(to, item.address, chain, destChain);
-                  }}
-                  renderSelected={() => (
-                    <AccountSelectField
-                      label={i18n.inputLabel.sendFrom}
-                      accountName={senderAccountName}
-                      value={from}
-                      showIcon
-                      outerStyle={{ marginBottom: 8 }}
-                    />
-                  )}
+              <View style={stylesheet.subheader}>
+                <SubHeader
+                  title={viewStep === 1 ? title : i18n.common.amount}
+                  onPressBack={onSubheaderPressBack}
                   disabled={loading}
-                  accountSelectorRef={accountSelectorRef}
+                  titleTextAlign={'left'}
                 />
-              )}
-
-              <View style={{ flexDirection: 'row', gap: theme.sizeSM, paddingBottom: theme.sizeXXS }}>
-                <View style={{ flex: 1 }}>
-                  <TokenSelector
-                    items={tokenItems}
-                    selectedValueMap={{ [asset]: true }}
-                    onSelectItem={item => {
-                      onChangeAssetValue(item.slug);
-                      onChangeValue('destChain')(item.originChain);
-                      tokenSelectorRef && tokenSelectorRef.current?.onCloseModal();
-                      setIsTransferAll(false);
-                      setForceUpdateMaxValue(undefined);
-                      isToAddressDirty && validateRecipientAddress(to, from, item.originChain, item.originChain);
-                    }}
-                    tokenSelectorRef={tokenSelectorRef}
-                    renderSelected={() => (
-                      <TokenSelectField
-                        logoKey={currentChainAsset?.symbol || ''}
-                        subLogoKey={currentChainAsset?.originChain || ''}
-                        value={currentChainAsset?.symbol || ''}
-                        showIcon
-                      />
-                    )}
-                    disabled={!tokenItems.length || loading}
-                  />
-                </View>
-
-                <View style={{ flex: 1 }}>
-                  <InputAmount
-                    forceUpdateMaxValue={forceUpdateMaxValue}
-                    disable={loading}
-                    value={amount}
-                    maxValue={maxTransfer}
-                    onChangeValue={_onChangeAmount}
-                    decimals={decimals}
-                    onSetMax={onSetMaxTransferable}
-                    showMaxButton
-                  />
-                </View>
               </View>
 
-              {!!(formState.errors.value && formState.errors.value.length) &&
-                formState.errors.value.map((message, index) => (
-                  <Warning key={index} isDanger message={message} style={{ marginBottom: 8 }} />
-                ))}
+              <>
+                <ScrollView
+                  style={stylesheet.scrollView}
+                  contentContainerStyle={stylesheet.scrollViewContentContainer}
+                  keyboardShouldPersistTaps={'handled'}>
+                  {isAllAccount && viewStep === 1 && (
+                    <>
+                      <AccountSelector
+                        items={accountItems}
+                        selectedValueMap={{ [fromValue]: true }}
+                        onSelectItem={_onChangeFrom}
+                        renderSelected={() => (
+                          <AccountSelectField
+                            label={i18n.inputLabel.sendFrom}
+                            accountName={senderAccountName}
+                            value={fromValue}
+                            showIcon
+                            outerStyle={{ marginBottom: theme.marginSM }}
+                          />
+                        )}
+                        disabled={loading}
+                        accountSelectorRef={accountSelectorRef}
+                      />
+                    </>
+                  )}
 
-              <InputAddress
-                ref={formState.refs.to}
-                containerStyle={{ marginBottom: theme.sizeSM }}
-                label={formState.labels.to}
-                value={formState.data.to}
-                onChangeText={onChangeRecipientAddress}
-                isValidValue={formState.isValidated.recipientAddress}
-                placeholder={i18n.placeholder.accountAddress}
-                disabled={loading}
-                onSubmitEditing={onSubmit}
-                addressPrefix={destChainNetworkPrefix}
-                networkGenesisHash={destChainGenesisHash}
-                showAddressBook
-                saveAddress
-              />
+                  <View style={stylesheet.row}>
+                    <View style={stylesheet.rowItem}>
+                      <TokenSelector
+                        items={viewStep === 1 ? tokenItems : tokenItemsViewStep2}
+                        selectedValueMap={{ [assetValue]: true }}
+                        onSelectItem={_onChangeAsset}
+                        showAddBtn={false}
+                        tokenSelectorRef={tokenSelectorRef}
+                        renderSelected={() => (
+                          <TokenSelectField
+                            logoKey={currentChainAsset?.slug || ''}
+                            subLogoKey={currentChainAsset?.originChain || ''}
+                            value={currentChainAsset?.symbol || ''}
+                            outerStyle={{ marginBottom: 0 }}
+                            showIcon
+                          />
+                        )}
+                        disabled={!tokenItems.length || loading}
+                      />
+                    </View>
 
-              {!!(formState.errors.to && formState.errors.to.length) &&
-                formState.errors.to.map((message, index) => (
-                  <Warning key={index} isDanger message={message} style={{ marginBottom: theme.marginSM }} />
-                ))}
+                    <View style={stylesheet.paperPlaneIconWrapper}>
+                      <Icon phosphorIcon={PaperPlaneRight} size={'md'} iconColor={theme['gray-5']} />
+                    </View>
 
-              <ChainSelector
-                items={destChainItems}
-                selectedValueMap={{ [destChain]: true }}
-                chainSelectorRef={chainSelectorRef}
-                onSelectItem={item => {
-                  onChangeValue('destChain')(item.slug);
-                  setForceUpdateMaxValue(isTransferAll ? {} : undefined);
-                  chainSelectorRef && chainSelectorRef.current?.onCloseModal();
-                  isToAddressDirty && validateRecipientAddress(to, from, chain, item.slug);
-                }}
-                renderSelected={() => (
-                  <NetworkField
-                    networkKey={destChain}
-                    outerStyle={{ marginBottom: 0 }}
-                    placeholder={i18n.placeholder.selectChain}
-                    showIcon
-                  />
-                )}
-                disabled={!destChainItems.length || loading}
-              />
-              <FreeBalance address={from} chain={chain} onBalanceReady={setIsBalanceReady} tokenSlug={asset} />
-            </ScrollView>
+                    <View style={stylesheet.rowItem}>
+                      <ChainSelector
+                        items={viewStep === 1 ? destChainItems : destChainItemsViewStep2}
+                        acceptDefaultValue={viewStep === 2 && destChainItemsViewStep2.length === 1}
+                        selectedValueMap={{ [destChainValue]: true }}
+                        chainSelectorRef={chainSelectorRef}
+                        onSelectItem={_onChangeDestChain}
+                        renderSelected={() => (
+                          <NetworkField
+                            networkKey={destChainValue}
+                            outerStyle={{ marginBottom: 0 }}
+                            placeholder={i18n.placeholder.selectChain}
+                            showIcon
+                          />
+                        )}
+                        disabled={!destChainItems.length || loading}
+                      />
+                    </View>
+                  </View>
 
-            <View
-              style={{
-                paddingHorizontal: 16,
-                paddingTop: 16,
-                ...MarginBottomForSubmitButton,
-              }}>
-              <Button
-                disabled={
-                  isDataNotReady ||
-                  !isBalanceReady ||
-                  !!formState.errors.to.length ||
-                  !!formState.errors.value.length ||
-                  loading
-                }
-                icon={buttonIcon}
-                loading={loading}
-                type={isTransferAll ? 'warning' : undefined}
-                onPress={onSubmit}>
-                {isTransferAll ? i18n.buttonTitles.transferAll : i18n.buttonTitles.transfer}
-              </Button>
-            </View>
-            <SafeAreaView />
-          </>
-        </>
-      </ScreenContainer>
-    </KeyboardAvoidingView>
+                  {viewStep === 1 && (
+                    <>
+                      <FormItem
+                        control={control}
+                        rules={recipientAddressRules}
+                        render={({ field: { value, ref, onChange, onBlur } }) => (
+                          <InputAddress
+                            ref={ref}
+                            label={i18n.inputLabel.sendTo}
+                            value={value}
+                            onChangeText={onChange}
+                            onBlur={onBlur}
+                            reValidate={reValidate}
+                            onSideEffectChange={onBlur}
+                            placeholder={i18n.placeholder.accountAddress}
+                            disabled={loading}
+                            addressPrefix={destChainNetworkPrefix}
+                            networkGenesisHash={destChainGenesisHash}
+                            chain={chainValue}
+                            showAddressBook
+                            saveAddress
+                          />
+                        )}
+                        name="to"
+                      />
+                    </>
+                  )}
+
+                  {viewStep === 2 ? (
+                    <View style={stylesheet.amountWrapper}>
+                      <FormItem control={control} rules={amountRules} render={renderAmountInput} name="value" />
+                    </View>
+                  ) : (
+                    <View style={stylesheet.balanceWrapper}>
+                      {!(!fromValue && !chainValue) && (
+                        <FreeBalanceDisplay
+                          tokenSlug={assetValue}
+                          nativeTokenBalance={nativeTokenBalance}
+                          nativeTokenSlug={nativeTokenSlug}
+                          tokenBalance={tokenBalance}
+                          style={stylesheet.balance}
+                          error={isGetBalanceError}
+                          isLoading={isGetBalanceLoading}
+                        />
+                      )}
+                    </View>
+                  )}
+                </ScrollView>
+
+                <View style={stylesheet.footer}>
+                  {viewStep === 1 && (
+                    <Button
+                      disabled={isNextButtonDisable}
+                      icon={getButtonIcon(ArrowCircleRight)}
+                      onPress={() => {
+                        trigger('to').then(pass => {
+                          if (pass) {
+                            setViewStep(2);
+                          }
+                        });
+                      }}>
+                      {i18n.buttonTitles.next}
+                    </Button>
+                  )}
+                  {viewStep === 2 && (
+                    <>
+                      <View style={stylesheet.footerBalanceWrapper}>
+                        <FreeBalanceDisplay
+                          label={
+                            viewStep === 2
+                              ? `${
+                                  i18n.customization.balance[0].toUpperCase() +
+                                  i18n.customization.balance.slice(1).toLowerCase()
+                                }:`
+                              : undefined
+                          }
+                          tokenSlug={assetValue}
+                          nativeTokenBalance={nativeTokenBalance}
+                          nativeTokenSlug={nativeTokenSlug}
+                          tokenBalance={tokenBalance}
+                          style={stylesheet.balance}
+                          error={isGetBalanceError}
+                          isLoading={isGetBalanceLoading}
+                        />
+
+                        {viewStep === 2 && (
+                          <TouchableOpacity
+                            onPress={() => {
+                              setForceUpdateValue({ value: maxTransfer });
+                              const bnMaxTransfer = new BN(maxTransfer);
+
+                              if (!bnMaxTransfer.isZero()) {
+                                setIsTransferAll(true);
+                              }
+                            }}
+                            style={stylesheet.max}>
+                            {<Typography.Text style={stylesheet.maxText}>{i18n.common.max}</Typography.Text>}
+                          </TouchableOpacity>
+                        )}
+                      </View>
+                      <Button
+                        disabled={isSubmitButtonDisable}
+                        loading={loading}
+                        type={isTransferAll ? 'warning' : undefined}
+                        onPress={checkAction(handleSubmit(onSubmit), extrinsicType)}
+                        icon={getButtonIcon(PaperPlaneTilt)}>
+                        {isTransferAll ? i18n.buttonTitles.transferAll : i18n.buttonTitles.transfer}
+                      </Button>
+                    </>
+                  )}
+                </View>
+                <SafeAreaView />
+              </>
+            </>
+          </ScreenContainer>
+        </KeyboardAvoidingView>
+      ) : (
+        <TransactionDone transactionDoneInfo={transactionDoneInfo} />
+      )}
+    </>
   );
 };
