+++ src/screens/Transaction/SendFund/styles/Amount.ts
